+++ src/checkbox-select-all/index.js
@@ -3,7 +3,7 @@ import { Controller } from 'stimulus'
 export default class extends Controller {
   static targets = ['checkbox']
 
-  toggleAll (e) {
+  toggle (e) {
     e.preventDefault()
 
     this.checkboxTargets.forEach(checkbox => {