+++ src/index.js
@@ -2,5 +2,6 @@ import Reveal from './reveal'
 import NestedForm from './nested-form'
 import CheckboxSelectAll from './checkbox-select-all'
 import Remote from './remote'
+import ContentLoader from './content-loader'
 
-export { Reveal, NestedForm, CheckboxSelectAll, Remote }
+export { Reveal, NestedForm, CheckboxSelectAll, Remote, ContentLoader }