+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Adding [content-loader controller](https://github.com/guillaumebriday/stimulus-components/tree/master/src/content-loader)
+
 ### Change
 
 - **Breaking**: Using data-maps in [reveal controller](https://github.com/guillaumebriday/stimulus-components/tree/master/src/reveal) instead of plain data attribute for hidden class.