+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2020-10-14
+
 ### Added
 
 - Adding [content-loader controller](https://github.com/guillaumebriday/stimulus-components/tree/master/src/content-loader)